chore: file structure and types
